+++ include/DijkstraConstrained.h
@@ -18,25 +18,25 @@ namespace Graphs
   class DijkstraConstrained : Algorithm
     {
         private:
-          Node              *startNode;
-	  float             maxCost;
+            Node                *startNode;
+	        float               maxCost;
 
-          vector<int>     distances;
-          vector<Node*>     fromNode;
-	  vector<Node*>     reachedNodes;
+            vector<int>         distances;
+            vector<Node*>       fromNode;
+	        vector<Node*>       reachedNodes;
 
         public:
-          DijkstraConstrained(Graph *graph, int startNode, float maxCost);
+            DijkstraConstrained(Graph *graph, int startNode, int maxCost);
 
-          bool              execute() override;
-          bool              execute(int startNode);
-	  bool              execute(int startNode, float maxCost);
+            bool                execute() override;
+            bool                execute(int startNode);
+	        bool                execute(int startNode, int maxCost);
 
-          vector<Node*>     getNodesOnShortestPathTo(int node);
-	  vector<Node*>     getReachedNodes();
-	  bool              isReached(int node);
-          vector<Edge*>     getEdgesOnShortestPathTo(int node);
-          float             getCostTo(int node);
+            vector<Node*>       getNodesOnShortestPathTo(int node);
+	        vector<Node*>       getReachedNodes();
+	        bool                isReached(int node);
+            vector<Edge*>       getEdgesOnShortestPathTo(int node);
+            int                 getCostTo(int node);
     };
 
 }
